Split SkiaChartRenderer
